Rename baseController to driverFactory (controller / tasksManager side changes)
Note: while renaming noticed newTestTasksManager() was in the wrong file and
moved to tasks_manager_test.go
